refactor: cleanup
[skip ci]

+++ src/apu/channels/mod.rs
@@ -13,10 +13,10 @@ pub trait Channel {
     fn is_mute(&self) -> bool;
     fn set_enabled(&mut self, value: bool);
 
-    fn get_sample(&self) -> u8 {
+    fn get_sample(&self) -> f32 {
         match self.is_mute() {
-            true => 0,
-            false => self.raw_sample(),
+            true => 0.0,
+            false => self.raw_sample() as f32,
         }
     }
 }
